review: code review & fix;

+++ packages/cli/src/actions/function.ts
@@ -1,12 +1,13 @@
+import { PROJECT_DIR } from './../utils/constants'
 
-import  * as path  from 'node:path'
+import * as path from 'node:path'
 import * as fs from 'node:fs'
 import { compileTs2js } from '../utils/util-lang'
 
-import { FUNCTIONS_DIR ,FUNCTIONS_FILE} from '../utils/constants'
-import { checkDir} from '../utils/util'
+import { FUNCTIONS_DIR, FUNCTIONS_FILE } from '../utils/constants'
+import { ensureDirectory } from '../utils/util'
 
-import { debugFunction,getFunctionByName,pushFunction,createFunction} from '../api/functions'
+import { debugFunction, getFunctionByName, pushFunction, createFunction } from '../api/functions'
 
 
 
@@ -17,45 +18,47 @@ import { debugFunction,getFunctionByName,pushFunction,createFunction} from '../a
  * @returns
  */
 
-export async function handlePullFunctionCommand(data:any,options:any) {
+export async function handlePullFunctionCommand(data: any, options: any) {
 
     // functions dir
-    const functionsDir = path.resolve(process.cwd(), FUNCTIONS_DIR)
+    const functionsDir = path.resolve(PROJECT_DIR, FUNCTIONS_DIR)
 
-    checkDir(functionsDir)
+    ensureDirectory(functionsDir)
 
     data.forEach(element => {
 
         //fuction name
-        const funcName =element.name;
+        const funcName = element.name
         const funcNameDir = path.resolve(functionsDir, funcName)
+        ensureDirectory(funcNameDir)
 
-        checkDir(funcNameDir)
+        const funcFile = path.resolve(funcNameDir, FUNCTIONS_FILE)
 
-        const funcFile= path.resolve(funcNameDir, FUNCTIONS_FILE)
-        try{
-            // check if exist function file
-            fs.accessSync(funcFile)
-            const currentCode =fs.readFileSync(funcFile,'utf-8')
-            
-            if(currentCode){
-                // forceOverwrite
-                if(options.forceOverwrite){
-                    fs.writeFileSync(funcFile, element.code)
-                }
-            }else{
-                fs.writeFileSync(funcFile, element.code)
-            }
-        }catch(err){
-
+        // create function file if not exists
+        const file_existed = fs.existsSync(funcFile)
+        if (!file_existed) {
             fs.writeFileSync(funcFile, element.code)
-
+            console.log(`${funcName} created`)
+            return
         }
-        
-        console.log('pull success')
-    })
 
-    
+        // force update function file if force option is true
+        const currentCode = fs.readFileSync(funcFile, 'utf-8')
+        if (options.forceOverwrite) {
+            fs.writeFileSync(funcFile, element.code)
+            console.log(`${funcName} updated`)
+            return
+        }
+
+        // update function file if code is empty
+        if (!currentCode) {
+            fs.writeFileSync(funcFile, element.code)
+            console.log(`${funcName} updated`)
+            return
+        }
+
+        console.log(`${funcName} unchanged`)
+    })
 }
 
 /**
@@ -66,26 +69,24 @@ export async function handlePullFunctionCommand(data:any,options:any) {
  * @returns
  */
 
-export async function handleInvokeFunctionCommand(appid:string,functionName:string,param:object) {
-
-    const functionsDir = path.resolve(process.cwd(), FUNCTIONS_DIR)
+export async function handleInvokeFunctionCommand(appid: string, functionName: string, param: object) {
 
     // get local code
-    const functionNameDir = path.resolve(functionsDir, functionName)
-    const funcFile= path.resolve(functionNameDir, FUNCTIONS_FILE)
+    const functionNameDir = path.resolve(PROJECT_DIR, FUNCTIONS_DIR, functionName)
+    const funcFile = path.resolve(functionNameDir, FUNCTIONS_FILE)
     const code = fs.readFileSync(funcFile, 'utf8')
     const obj = {
-        func:{
+        func: {
             appid: appid,
             code: code,
-            name:functionName,
+            name: functionName,
             compiledCode: compileTs2js(code),
             debugParams: JSON.stringify(param),
         },
-        param:param
+        param: param
     }
 
-    const res = await debugFunction(functionName,obj)
+    const res = await debugFunction(functionName, obj)
     console.log(res)
 
 }
@@ -98,54 +99,46 @@ export async function handleInvokeFunctionCommand(appid:string,functionName:stri
  * @param {any} options
  * @returns
  */
+export async function handlePushFunctionCommand(appid: string, functionName: string, options: any) {
 
-
-export async function handlePushFunctionCommand(appid:string,functionName:string,options:any) {
-
-    const functionsDir = path.resolve(process.cwd(), FUNCTIONS_DIR)
+    const functionsDir = path.resolve(PROJECT_DIR, FUNCTIONS_DIR)
 
     // get local code
     const functionNameDir = path.resolve(functionsDir, functionName)
-    const funcFile= path.resolve(functionNameDir, FUNCTIONS_FILE)
-    const  code = fs.readFileSync(funcFile, 'utf8')
-    
-    // get function
-    const record = await getFunctionByName(appid,functionName)
-    
-    //update function
-    if(record.data){
-        if(record.data.code!==code){
-    
-            if(options.forceOverwrite){
-                const data = {
-                    code:code,
-                    debugParams:JSON.stringify({"code":"laf"}),
-                }
-                const res = await pushFunction(appid,functionName,data)
-                if(res.data){
-                    console.log("push success")
-                }
-            }else{
+    const funcFile = path.resolve(functionNameDir, FUNCTIONS_FILE)
+    const code = fs.readFileSync(funcFile, 'utf8')
 
-                console.log("romote code is different with local")
-            }
-        }else{
-            console.log("romote code is same with local")
-        }
-    
-    }else{
+    // get remote function
+    const record = await getFunctionByName(appid, functionName)
+
+    // create function if not exists
+    if (!record.data) {
         // create function
         const data = {
-            code:code,
-            name:functionName,
-            label:"test",
-            status:1
+            code: code,
+            name: functionName,
+            label: functionName,
+            status: 1,
+            compiledCode: compileTs2js(code),
+            debugParams: '{}',
         }
-    
-        const res = await createFunction(appid,data)
-        if(res.data){
+
+        const res = await createFunction(appid, data)
+        if (res.data) {
             console.log("push success")
-        }     
+        }
+        return
     }
-    
+
+    // update function if code is different
+    if (record.data.code !== code) {
+        const data = {
+            code: code,
+            debugParams: JSON.stringify({ "code": "laf" }),
+        }
+        const res = await pushFunction(appid, functionName, data)
+        if (res.data) {
+            console.log("push success")
+        }
+    }
 }
